@@ -4,30 +4,26 @@
   <groupId>com.yapicimurat</groupId>
   <artifactId>mywebsite</artifactId>
   <packaging>jar</packaging>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>11</source>
-                    <target>11</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-    <version>1.0-SNAPSHOT</version>
+  <build>
+      <plugins>
+          <plugin>
+              <groupId>org.apache.maven.plugins</groupId>
+              <artifactId>maven-compiler-plugin</artifactId>
+              <configuration>
+                  <source>${maven.compiler.source}</source>
+                  <target>${maven.compiler.target}</target>
+              </configuration>
+          </plugin>
+      </plugins>
+  </build>
+  <version>1.0-SNAPSHOT</version>
   <name>mywebsite</name>
   <url>http://maven.apache.org</url>
-
   <properties>
+    <java.version>22</java.version>
     <org.mapstruct.version>1.5.5.Final</org.mapstruct.version>
-    <!-- Maven compiler properties to read-->
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-
-    <!-- General java version properties defining to use from other mavens...-->
-    <java.version>11</java.version>
+    <maven.compiler.source>22</maven.compiler.source>
+    <maven.compiler.target>22</maven.compiler.target>
   </properties>
 
   <parent>
@@ -50,9 +46,6 @@
       <artifactId>mapstruct</artifactId>
       <version>${org.mapstruct.version}</version>
     </dependency>
-
-
-
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
@@ -80,17 +73,10 @@
       <artifactId>mysql-connector-j</artifactId>
       <version>8.0.31</version>
     </dependency>
-
-    <!-- https://mvnrepository.com/artifact/org.modelmapper/modelmapper -->
-    <dependency>
-      <groupId>org.modelmapper</groupId>
-      <artifactId>modelmapper</artifactId>
-      <version>3.0.0</version>
-    </dependency>
-
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-validation</artifactId>
+      <version>2.7.6</version>
     </dependency>
 
   </dependencies>
