@@ -1,17 +1,12 @@
 package com.yapicimurat;
 
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class App implements CommandLineRunner {
+public class App {
     public static void main( String[] args )
     {
         SpringApplication.run(App.class, args);
-    }
-    @Override
-    public void run(String... args) throws Exception {
-
     }
 }
