@@ -3,14 +3,14 @@ package com.yapicimurat.repository;
 import com.yapicimurat.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
 @Repository
 public interface CategoryRepository extends JpaRepository<Category, UUID> {
-    Set<Category> findAllByIdIn(Set<UUID> ids);
+    List<Category> findAllByIdIn(Set<UUID> ids);
     boolean existsByName(String name);
-    Set<Category> findAllById(UUID id);
-    long countAllByIdIn(Set<UUID> ids);
-    long countAllByIdInAndNameIn(Set<UUID> ids, Set<String> name);
+    boolean existsByIdNotAndName(UUID id, String name);
+    long countByNameIn(Set<String> nameSet);
 }
