@@ -1,22 +1,10 @@
-package com.yapicimurat.dto;
+package com.yapicimurat.web.input;
 
-import java.util.List;
-
-public class CommentDTO {
-    private String id;
+public class CommentUpdateRequest {
     private String name;
     private String lastName;
     private String email;
     private String text;
-    private Boolean isTest;
-    private CommentDTO parentComment;
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public String getName() {
         return name;
@@ -49,22 +37,4 @@ public class CommentDTO {
     public void setText(String text) {
         this.text = text;
     }
-
-
-    public Boolean getTest() {
-        return isTest;
-    }
-
-    public void setTest(Boolean test) {
-        isTest = test;
-    }
-
-    public CommentDTO getParentComment() {
-        return parentComment;
-    }
-
-    public void setParentComment(CommentDTO parentComment) {
-        this.parentComment = parentComment;
-    }
-
 }
